+++ src/test/java/com/arup/cml/abm/kpi/integration/MatsimKpiGeneratorIntegrationTest.java
@@ -4,6 +4,8 @@ import com.arup.cml.abm.kpi.matsim.run.MatsimKpiGenerator;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
+import org.matsim.core.utils.misc.CRCChecksum;
+
 import picocli.CommandLine;
 
 import java.io.File;
@@ -13,6 +15,7 @@ import static java.lang.String.format;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 public class MatsimKpiGeneratorIntegrationTest {
+    private static final String COMPRESSION_FILE_EXTENSION = ".gz";
 
     @Rule
     public TemporaryFolder appOutputDir = new TemporaryFolder();
@@ -36,22 +39,41 @@ public class MatsimKpiGeneratorIntegrationTest {
         assertSupportingFilesWereGenerated(appOutputDir.getRoot());
     }
 
+    @Test
+    public void testAppWithDrt() throws Exception {
+        String testDataDirRoot = format("%s/integration-test-data/drt-matsim-outputs/",
+                Paths.get("src", "test", "resources"));
+
+        int exitCode = new CommandLine(new MatsimKpiGenerator()).execute(
+                "-mc",
+                format("%s/output_config.xml", testDataDirRoot),
+                "-mo",
+                testDataDirRoot,
+                "-o",
+                appOutputDir.getRoot().getAbsolutePath()
+        );
+
+        assertThat(exitCode).isEqualTo(0).as("App return code should be zero");
+        assertKpiFilesWereGenerated(format("%s/expected-kpis", testDataDirRoot), appOutputDir.getRoot());
+        assertSupportingFilesWereGenerated(appOutputDir.getRoot());
+    }
+
     private void assertSupportingFilesWereGenerated(File kpiDirectory) {
         String[] generatedFiles = kpiDirectory.list();
-        String [] expectedSupportingFiles = {
-                "intermediate-pt-wait-time.csv",
-                "intermediate-occupancy-rate.csv",
-                "intermediate-congestion.csv",
-                "intermediate-vehicle-km.csv",
-                "supporting-data-vehicles.csv",
-                "supporting-data-scheduleRoutes.csv",
-                "supporting-data-scheduleStops.csv",
-                "supporting-data-networkLinkModes.csv",
-                "supporting-data-networkLinks.csv",
-                "supporting-data-vehicleOccupancy.csv",
-                "supporting-data-linkLog.csv",
-                "supporting-data-trips.csv",
-                "supporting-data-legs.csv"
+        String[] expectedSupportingFiles = {
+                "intermediate-pt-wait-time.csv" + COMPRESSION_FILE_EXTENSION,
+                "intermediate-occupancy-rate.csv" + COMPRESSION_FILE_EXTENSION,
+                "intermediate-congestion.csv" + COMPRESSION_FILE_EXTENSION,
+                "intermediate-vehicle-km.csv" + COMPRESSION_FILE_EXTENSION,
+                "supporting-data-vehicles.csv" + COMPRESSION_FILE_EXTENSION,
+                "supporting-data-scheduleRoutes.csv" + COMPRESSION_FILE_EXTENSION,
+                "supporting-data-scheduleStops.csv" + COMPRESSION_FILE_EXTENSION,
+                "supporting-data-networkLinkModes.csv" + COMPRESSION_FILE_EXTENSION,
+                "supporting-data-networkLinks.csv" + COMPRESSION_FILE_EXTENSION,
+                "supporting-data-vehicleOccupancy.csv" + COMPRESSION_FILE_EXTENSION,
+                "supporting-data-linkLog.csv" + COMPRESSION_FILE_EXTENSION,
+                "supporting-data-trips.csv" + COMPRESSION_FILE_EXTENSION,
+                "supporting-data-legs.csv" + COMPRESSION_FILE_EXTENSION
         };
         for (int i = 0; i < expectedSupportingFiles.length; i++) {
             assertThat(generatedFiles)
@@ -62,13 +84,13 @@ public class MatsimKpiGeneratorIntegrationTest {
 
     private void assertKpiFilesWereGenerated(String expectedKpiDirectory, File kpiDirectory) {
         String[] generatedFiles = kpiDirectory.list();
-        String [] expectedKpiFiles = {
-                "kpi-congestion.csv",
-                "kpi-speed.csv",
-                "kpi-vehicle-km.csv",
-                "kpi-occupancy-rate.csv",
-                "kpi-modal-split.csv",
-                "kpi-pt-wait-time.csv",
+        String[] expectedKpiFiles = {
+                "kpi-congestion.csv" + COMPRESSION_FILE_EXTENSION,
+                "kpi-speed.csv" + COMPRESSION_FILE_EXTENSION,
+                "kpi-vehicle-km.csv" + COMPRESSION_FILE_EXTENSION,
+                "kpi-occupancy-rate.csv" + COMPRESSION_FILE_EXTENSION,
+                "kpi-modal-split.csv" + COMPRESSION_FILE_EXTENSION,
+                "kpi-pt-wait-time.csv" + COMPRESSION_FILE_EXTENSION,
         };
         for (int i = 0; i < expectedKpiFiles.length; i++) {
             String kpiFile = expectedKpiFiles[i];
@@ -76,9 +98,10 @@ public class MatsimKpiGeneratorIntegrationTest {
                     .contains(kpiFile)
                     .as(format("Check KPI output file '%s' exists", kpiFile));
             File expectedKpiFile = new File(format("%s/expected-%s", expectedKpiDirectory, kpiFile));
-            assertThat(new File(format("%s/%s", kpiDirectory, kpiFile)))
-                    .hasSameTextualContentAs(expectedKpiFile)
-                    .as(format("Check %s KPI data matches expectation", kpiFile));
+            File currentKpiFile = new File(format("%s/%s", kpiDirectory, kpiFile));
+            long expected = CRCChecksum.getCRCFromFile(expectedKpiFile.toString());
+            long current = CRCChecksum.getCRCFromFile(currentKpiFile.toString());
+            assertThat(expected).isEqualTo(current).as(format("Check %s KPI data matches expectation", kpiFile));
         }
     }
 }